Update changes
- Core features in Home panel.
- Some settings, but not usable yet.
- Pull down to refresh

+++ app/src/main/java/io/zoemeow/dutapp/model/subject/SubjectFeeListItem.kt
@@ -3,6 +3,6 @@ package io.zoemeow.dutapp.model.subject
 data class SubjectFeeListItem(
     val date: Long? = null,
     val total_credit: Int? = null,
-    val total_money: Int? = null,
+    val total_money: Long? = null,
     val fee_list: ArrayList<SubjectFeeItem>? = null
 )
